+++ routes/actions.php
@@ -1,0 +1,6 @@
+<?php
+
+use Illuminate\Support\Facades\Route;
+use TransformStudios\Front\Http\Controllers\ConfigController;
+
+Route::get('config', ConfigController::class)->name('front.config');
